KOGITO-7954 : SWF Editor - Some basic automated integration tests
