+++ packages/serverless-workflow-diagram-editor/sw-editor/sw-editor-kogito-app/src/test/java/org/kie/workbench/common/stunner/sw/client/selenium/SWAPITests.java
@@ -16,11 +16,9 @@
 
 package org.kie.workbench.common.stunner.sw.client.selenium;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.Test;
-import org.openqa.selenium.JavascriptExecutor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -32,14 +30,12 @@ public class SWAPITests extends SWEditorSeleniumBase {
         setContent(expected);
         waitCanvasPanel();
 
-        JsCanvasHelper jsHelper = new JsCanvasHelper((JavascriptExecutor) driver);
         List<String> nodeIds = jsHelper.getNodeIds();
         assertThat(nodeIds.size()).isEqualTo(2);
 
         for (String uuid : nodeIds) {
             String backgroundColor = jsHelper.getGetBackgroundColor(uuid);
             String borderColor = jsHelper.getBorderColor(uuid);
-            ;
 
             assertThat(backgroundColor).isEqualTo("#ffffff");
             assertThat(borderColor).isEqualTo("#d5d5d5");
@@ -62,7 +58,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
         setContent(expected);
         waitCanvasPanel();
 
-        JsCanvasHelper jsHelper = new JsCanvasHelper((JavascriptExecutor) driver);
         List<String> nodeIds = jsHelper.getNodeIds();
         assertThat(nodeIds.size()).isEqualTo(2);
 
@@ -93,7 +88,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
         setContent(expected);
         waitCanvasPanel();
 
-        JsCanvasHelper jsHelper = new JsCanvasHelper((JavascriptExecutor) driver);
         List<String> nodeIds = jsHelper.getNodeIds();
         assertThat(nodeIds.size()).isEqualTo(2);
 
@@ -114,7 +108,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
         setContent(expected);
         waitCanvasPanel();
 
-        JsCanvasHelper jsHelper = new JsCanvasHelper((JavascriptExecutor) driver);
         List<String> nodeIds = jsHelper.getNodeIds();
         assertThat(nodeIds.size()).isEqualTo(2);
 
@@ -128,7 +121,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
 
         /// Invalid
         jsHelper.applyState(uuid, "invalid");
-        Thread.sleep(2000); // Otherwise sometime it applies the state and sometimes it does not finish
 
         backgroundColor = jsHelper.getGetBackgroundColor(uuid);
         borderColor = jsHelper.getBorderColor(uuid);
@@ -138,7 +130,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
 
         /// Highlight
         jsHelper.applyState(uuid, "highlight");
-        Thread.sleep(2000); // Otherwise sometime it applies the state and sometimes it does not finish
 
         backgroundColor = jsHelper.getGetBackgroundColor(uuid);
         borderColor = jsHelper.getBorderColor(uuid);
@@ -148,7 +139,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
 
         /// Selected
         jsHelper.applyState(uuid, "selected");
-        Thread.sleep(2000); // Otherwise sometime it applies the state and sometimes it does not finish
 
         backgroundColor = jsHelper.getGetBackgroundColor(uuid);
         borderColor = jsHelper.getBorderColor(uuid);
@@ -158,7 +148,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
 
         /// None
         jsHelper.applyState(uuid, "none");
-        Thread.sleep(2000); // Otherwise sometime it applies the state and sometimes it does not finish
 
         backgroundColor = jsHelper.getGetBackgroundColor(uuid);
         borderColor = jsHelper.getBorderColor(uuid);
@@ -174,7 +163,6 @@ public class SWAPITests extends SWEditorSeleniumBase {
         setContent(expected);
         waitCanvasPanel();
 
-        JsCanvasHelper jsHelper = new JsCanvasHelper((JavascriptExecutor) driver);
         List<String> nodeIds = jsHelper.getNodeIds();
 
         assertThat(nodeIds.size()).isEqualTo(2);
@@ -198,93 +186,10 @@ public class SWAPITests extends SWEditorSeleniumBase {
         setContent(expected);
         waitCanvasPanel();
 
-        JsCanvasHelper jsHelper = new JsCanvasHelper((JavascriptExecutor) driver);
         List<String> nodeIds = jsHelper.getNodeIds();
         assertThat(nodeIds.size()).isEqualTo(2);
 
         Boolean isConnection = jsHelper.isConnection(nodeIds.get(0), nodeIds.get(1));
         assertThat(isConnection).isEqualTo(true);
-    }
-
-    public class JsCanvasHelper {
-
-        private static final String GET_NODE_IDS = "return canvas.getNodeIds();";
-        private static final String GET_BACKGROUND_COLOR = "return canvas.getBackgroundColor('%s')";
-        private static final String GET_BORDER_COLOR = "return canvas.getBorderColor('%s')";
-        private static final String SET_BACKGROUND_COLOR = "return canvas.setBackgroundColor('%1','%2');";
-        private static final String SET_BORDER_COLOR = "return canvas.setBorderColor('%1', '%2');";
-        private static final String GET_LOCATION = "return canvas.getLocation('%s');";
-        private static final String GET_ABSOLUTE_LOCATION = "return canvas.getAbsoluteLocation('%s')";
-        private static final String GET_DIMENSIONS = "return canvas.getDimensions('%s')";
-        private static final String APPLY_STATE = "return canvas.applyState('%1','%2');";
-        private static final String CENTER_NODE = "return canvas.centerNode('%s');";
-        private static final String IS_CONNECTION = "return canvas.isConnected('%1', '%2');";
-
-        private JavascriptExecutor executor;
-
-        public JsCanvasHelper(JavascriptExecutor executor) {
-            this.executor = executor;
-        }
-
-        public List<String> getNodeIds() {
-            final Object result = executor.executeScript(String.format(GET_NODE_IDS));
-            assertThat(result).isInstanceOf(ArrayList.class);
-            return (ArrayList<String>) result;
-        }
-
-        public String getBorderColor(String uuid) {
-            Object borderColorResult = executor.executeScript(String.format(GET_BORDER_COLOR.replace("%s", uuid)));
-            assertThat(borderColorResult).isInstanceOf(String.class);
-            return (String) borderColorResult;
-        }
-
-        public String getGetBackgroundColor(String uuid) {
-            Object backgroundColorResult = executor.executeScript(String.format(GET_BACKGROUND_COLOR.replace("%s", uuid)));
-            assertThat(backgroundColorResult).isInstanceOf(String.class);
-            return (String) backgroundColorResult;
-        }
-
-        public void setBackgroundColor(String uuid, String color) {
-            String command = SET_BACKGROUND_COLOR.replace("%1", uuid).replace("%2", color);
-            executor.executeScript(String.format(command));
-        }
-
-        public void setBorderColor(String uuid, String color) {
-            String command = SET_BORDER_COLOR.replace("%1", uuid).replace("%2", color);
-            executor.executeScript(String.format(command));
-        }
-
-        public List<Object> getLocation(String uuid) {
-            Object location = executor.executeScript(String.format(GET_LOCATION.replace("%s", uuid)));
-            assertThat(location).isInstanceOf(ArrayList.class);
-            return (ArrayList<Object>) location;
-        }
-
-        public List<Object> getAbsoluteLocation(String uuid) {
-            Object absoluteLocation = executor.executeScript(String.format(GET_ABSOLUTE_LOCATION.replace("%s", uuid)));
-            assertThat(absoluteLocation).isInstanceOf(ArrayList.class);
-            return (ArrayList<Object>) absoluteLocation;
-        }
-
-        public List<Object> getDimension(String uuid) {
-            Object dimension = ((JavascriptExecutor) driver).executeScript(String.format(GET_DIMENSIONS.replace("%s", uuid)));
-            assertThat(dimension).isInstanceOf(ArrayList.class);
-            return (ArrayList<Object>) dimension;
-        }
-
-        public void applyState(String uuid, String state) {
-            String command = APPLY_STATE.replace("%1", uuid).replace("%2", state);
-            executor.executeScript(String.format(command));
-        }
-
-        public void centerNode(String uuid) {
-            executor.executeScript(String.format(CENTER_NODE.replace("%s", uuid)));
-        }
-
-        public boolean isConnection(String uuid1, String uuid2) {
-            final Object connection = ((JavascriptExecutor) driver).executeScript(String.format(IS_CONNECTION.replace("%1", uuid1).replace("%2", uuid2)));
-            assertThat(connection).isInstanceOf(Boolean.class);
-            return (Boolean) connection;
-        }
     }
 }
